1/functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.26.1
azure-data-tables==12.4.1
azure-functions==1.12.0
certifi==2022.9.24
charset-normalizer==2.1.1
DateTime==4.7
idna==3.4
isodate==0.6.1
msrest==0.7.1
multidict==6.0.2
oauthlib==3.2.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.12
uuid==1.30
yarl==1.8.1
zope.interface==5.5.1
